requirements.txt
asgiref==3.2.10
Django==3.0.5
django-extensions==3.0.9
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.2.4
uWSGI==2.0.19.1
